requirements.txt
Flask
line-bot-sdk
opencv-python
numpy
requests
